Re-apply changes
