+++ ws-server/src/lib.rs
@@ -26,13 +26,12 @@
 
 extern crate alloc;
 
-mod raw;
 mod server;
-mod transport;
+mod types;
 
-#[cfg(test)]
-mod tests;
+// #[cfg(test)]
+// mod tests;
 
-pub use raw::{RawServer as RawWsServer, RawServerEvent as RawWsServerEvent, TypedResponder as WsTypedResponder};
-pub use server::{RegisteredMethod, RegisteredNotification, Server as WsServer};
-pub use transport::WsTransportServer;
+// pub use server::{RegisteredMethod, RegisteredNotification, Server as WsServer};
+pub use server::Server as WsServer;
+// pub use transport::WsTransportServer;
